requirements.txt
# Core
google-genai>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
python-dotenv>=1.0.1
pydantic>=2.8.2
pydantic-settings>=2.4.0
rich>=13.0.0
aiosqlite>=0.19.0

# Dev
pytest>=8.0.0
pytest-asyncio>=0.23.0
